Remove loaded metamodel registering in the ResourceSet
If a metamodel is loaded, its registration in the global_registry or
the ResourceSet metamodel_registry is only required if an instance of
this metamodel is going to be loaded. In the case of the Ecore2Python
generation, the instance model is an Ecore instance metamodel. Thus,
the loaded metamodel is seen as an Ecore instance and it is not
required to register it in a metamodel_registry.