Fix missing templates in the distributed package and bump to 0.1.1
The 'package_data' in setuptools is quite hard to use
properly. Finally, I succeed add the missing templates, but the
solution was not so trivial. In facts, 'package_data' can only add
extra data that are contained in a package. If the data is not
contained in a package, the data is not added to the final
distribution build...