Add new 'kind' test for Jinja and modify templates accordingly
The introduction of the test 'kind' and the modification of the test
'type' tries to match the OCL semantic of 'oclIsKindOf' and
'oclIsTypeOf'. In this context, the differentiation is usefull for the
generation of EDataType and EEnum. EEnum inherits from EDataType and,
in some case, the code is equivalent (so 'kind' can be used) while in
other case, a strong difference is made in the generation (so 'type'
is used).