Add metamodel dependency generation with a new option
If the option is enabled, the generator also generates all the
metamodel that are 'dependencies' of the metamodel sets as input. A
metamodel dependency is typically a reference from the input metamodel
to another '.ecore'. This option introduces a slower code generation
as all metamodels must be scanned in order to extract the right
dependencies.