Add preliminary support for generic code gen
This addition proposes code generation for generic from .ecore models.
This generation is a first step, but does not tackle 'super', 'extend'
and other generic keywords. A part of the semantic is caught at runtime
and cannot be expressed in a pure classical statically typed way. The
result has minimal impact on performances and stays readable.